Allow indexing from backend index queue module
* Moved functionallity from scheduler task to IndexService to allow Indexing from different locations
* Added backend controller action to index a few documents from the backend

Fixes: #275
