+++ Resources/Private/Language/locallang.xlf
@@ -83,6 +83,10 @@
         <trans-unit id="solr.backend.index_queue_module.error_details" xml:space="preserve">
         	<source>Error details for queue item</source>
         </trans-unit>
+        <trans-unit id="solr.backend.index_queue_module.index_now" xml:space="preserve">
+        	<source>Index now!</source>
+        </trans-unit>
+
 		</body>
 	</file>
 </xliff>
